+++ crates/db/sqlite/src/lib.rs
@@ -10,7 +10,7 @@ extern crate alloc;
 pub mod agg;
 pub mod analyze;
 pub mod ast;
-mod bytes;
+pub mod bytes;
 pub mod change;
 pub mod check;
 pub mod constraint;