+++ crates/db/sqlite/src/tests/mod.rs
@@ -44,6 +44,7 @@ mod json;
 mod key_tree;
 mod matrix;
 mod misuse;
+mod newrowid;
 mod number;
 mod page;
 mod parse;